+++ relm4/src/shutdown/sender.rs
@@ -1,21 +1,22 @@
 // Copyright 2022 System76 <info@system76.com>
 // SPDX-License-Identifier: MIT or Apache-2.0
 
-use async_broadcast::Sender;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
+
+use tokio::sync::broadcast;
 
 /// Sends shutdown signals to receivers.
 #[derive(Debug)]
 pub struct ShutdownSender {
     pub(super) alive: Arc<AtomicBool>,
-    pub(super) sender: Sender<()>,
+    pub(super) sender: broadcast::Sender<()>,
 }
 
 impl ShutdownSender {
     /// Broadcasts a shutdown signal to listening receivers.
     pub fn shutdown(&self) {
-        drop(self.sender.broadcast(()));
+        drop(self.sender.send(()));
     }
 }